+++ lightsaber/src/main/java/schwarz/it/lightsaber/LightsaberSymbolProcessor.kt
@@ -5,38 +5,29 @@ import com.google.devtools.ksp.processing.Dependencies
 import com.google.devtools.ksp.processing.Resolver
 import com.google.devtools.ksp.processing.SymbolProcessor
 import com.google.devtools.ksp.symbol.KSAnnotated
-import com.google.devtools.ksp.symbol.KSFunctionDeclaration
-import com.google.devtools.ksp.symbol.KSType
-import dagger.Provides
+import schwarz.it.lightsaber.checkers.UnusedInject
 import java.io.PrintWriter
-import javax.inject.Inject
 
 class LightsaberSymbolProcessor(
     private val codeGenerator: CodeGenerator,
 ) : SymbolProcessor {
-    private val injects: MutableList<KSFunctionDeclaration> = mutableListOf()
-    private val provides: MutableSet<KSType> = mutableSetOf()
+    private val rules: Set<Pair<String, LightsaberSymbolProcessorRule>> = setOf(
+        "UnusedInject" to UnusedInject(),
+    )
 
     override fun process(resolver: Resolver): List<KSAnnotated> {
-        injects.addAll(
-            resolver.getSymbolsWithAnnotation(Inject::class.qualifiedName!!)
-                .filterIsInstance<KSFunctionDeclaration>(),
-        )
-        provides.addAll(
-            resolver.getSymbolsWithAnnotation(Provides::class.qualifiedName!!)
-                .filterIsInstance<KSFunctionDeclaration>()
-                .map { it.returnType!!.resolve() },
-        )
-
-        return emptyList()
+        return rules
+            .map { (_, rule) -> rule.process(resolver) }
+            .flatten()
     }
 
     override fun finish() {
-        val issues = injects
-            .filter { it.returnType!!.resolve() in provides }
+        val issues = rules
+            .map { (name, rule) ->
+                rule.computeFindings().map { Issue(it.codePosition, it.message, name) }
+            }
+            .flatten()
             .ifEmpty { return }
-            .map { Finding("This Inject is unused", it.location.toCodePosition()) }
-            .map { Issue(it.codePosition, it.message, "UnusedInject") }
 
         codeGenerator.createNewFile(Dependencies.ALL_FILES, "", "ksp", ".lightsaber")
             .let(::PrintWriter)
@@ -44,4 +35,10 @@ class LightsaberSymbolProcessor(
                 issues.forEach { writer.println(it.getMessage()) }
             }
     }
+}
+
+interface LightsaberSymbolProcessorRule {
+    fun process(resolver: Resolver): List<KSAnnotated>
+
+    fun computeFindings(): List<Finding>
 }
